+++ example/linux/flutter/generated_plugin_registrant.cc
@@ -2,8 +2,6 @@
 //  Generated file. Do not edit.
 //
 
-// clang-format off
-
 #include "generated_plugin_registrant.h"
 
 #include <flutter_local_authentication/flutter_local_authentication_plugin.h>